Use react router
Before this commit the state object determined what was shown on
screen. Using react router you can navigate to a screen even if it was
already filled in. This is needed if we want to be able to pick up were
we left of.

The new approach is also what the redux manual suggests.